httpkit/client: Only set the content-type if the body isnt empty
We should not send the Content-Type unless it's explicitly defined
through the spec. Currently the spec is not consulted for determining
the Content-Type so the client defaults to application/json. Not all
APIs can handle 'Content-Type: application/json' for every request, so
this change removes the content-type from requests without a body.

Still TODO: infer content-type from the spec/producer used.