Fix bug that caused unstable handlers' names after regeneration.
Before, if you generate a server, everything builds fine.  If you then delete
the models and restapi folders and regenerate, the code does not build.  This
is due to unstable naming of handlers' names that do not match the original
names in the server's configureAPI() function.

The fix is to modify the handler naming convention.  To guarantee uniqueness,
handler names are now a combination of method + path.