Adds copyright comment
Adds a copyright comment on top of auto-generated code files.
The copyright comment is fetched as a plain text from a file
input to swagger cli through -r option.
The copyright text is added to a /* */ block in each auto-generated go
file. Copyright must not contain any comment markers for golang (like */)