Fix model generation on discriminator conflict
Generating the model for a definition whose pascalized name matches
the discriminator results in a "pointer to an interface, not
interface" compiler error. To avoid this we compare the pascalized
version of the name to ensure that any characters that would be
removed in the pascalize transform are accounted for.

Signed-off-by: Christopher Grim <grim@fugue.co>