Edit the todo tutorial for grammar and clarity
Mostly this consisted of going through the tutorial and editing for English
grammar as well as trying to trim down the word count. The text was a bit more
verbose than it needed to be, which I think made it harder to follow.

I also made a few other changes:

* Wrapped more things in backticks to highlight technical terms, code
  snippets, etc.

* Updated output examples to match the current state of go-swagger (notably
  the generate files have changed a bit).