Make generate server sort the models #1349
The array used to generate the models is not currently sorted, so
the models are generated in whatever order comes out of the map.
When there are multiple models that fail to generate properly, this lack
of sorting results in changes in the reported failures from run to run.
Sorting results in deterministic behavior in the case of errors, making
it easier to fix the errors.

The generated operations are being sorted, this just applies the same
pattern to the models as well.

Signed-off-by: Greg Marr <greg.marr@autodesk.com>