Make format resolving strictly to require a valid pair of type and format.
The before implementation checks format only but it causes wrong type decisions.

e.g. {"type": "string", "format": "uint64"} had been converted to uint64.
     This behaivior is correct in case of "type" is "integer" or "number", but the "string" case is not.

This commit changes this decision algorithm. "format" is not treated as same of "type" anymore.
"format" must be available when it is specified with the valid (matched) "type". If not, the "format" will be ignored
and it just decides the type of Go code by using "type" value.

Signed-off-by: Kousuke Ebihara <kousuke@co3k.org>