Improve go.mod support in generator
This improves the `go.mod` support added in #1636 by adding support for
targets not at the module level.

Imagine you generate code like this:

```bash
$ swagger generate client --name my-api --target ./pkg/api
```

Then it would assume `go.mod` is under `pkg/api` and not detect that go
modules are used. The fix is to walk up the directory to find the first
occurance of `go.mod` and assume that as the module base URL.

Signed-off-by: Mikkel Oscar Lyderik Larsen <m@moscar.net>