Fix `goimports` step incorrectly rewriting imports
Before the fix `goimports` was run as if the generated file was to be written in
the directory where `swagger generate` was invoked.

Given the following directory structure:

./foo/client/operations/
./foo/client/foo_client.go
./quux/client/operations/
./quux/client/quux_client.go

`goimports` step for `quux/client/quux_client.go` changed import inside to
`foo/client/operations` due to a heuristic in `goimports` to prefer shorter
imports to longer ones.

Signed-off-by: Mikhail Gusarov <misha@ridge.co>