auth: added support for adding a principal as an interface
A new CLI flag --principal-is-interface is available to generate a server.
The principal will not be considered nullable and will generate authorizer methods without pointers.

* fixes #1659

* contributes to #2362

In #2362, it appeared that the stratoscale template was broken when using a custom principal.
This PR fixes this by generating correct code. However, users of this template will have to
provide the "configure_api" with a custom authorizer to handle this custom type, or the server will panic.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>