Configurable consumers in stratoscale templates
Currently the configuration code generated by the _stratoscale_
templates doesn't allow changing the consumers before creating the
handler. This makes it difficult to use custom consumers. For example,
we would like to use a JSON consumer that rejects unknown fields. To
make that easier this patch changes the templates so that the _Config_
object that is passed to the _Handler_ function has additional fields to
specify those custom consumers. With that in place we will be able to do
something like this:

```go
// myConsumer is a consumer function that rejects unknown fields.
func myConsumer() runtime.Consumer {
        return runtime.ConsumerFunc(func(reader io.Reader, data interface{}) error {
                dec := json.NewDecoder(reader)
                dec.UseNumber()
                dec.DisallowUnknownFields()
                return dec.Decode(data)
        })
}

// Create the handler using our custom consumer:
handler, err := restapi.Handler(restapi.Config{
	....
	JSONConsumer: myConsumer(),
})
```

Existing code will not be affected because the default value for this
new `JSONConsumer` field will be nil, and that results in the same
behaviour than before this change.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>