Fixes PropertiesSpecOrder not being used
Users can pass in --keep-spec-order to maintain the field order when
generating go code. This flag is not actually respected, and fields are
generated in alphabetical order. The reason for this is WithAutoXOrder
generates a new YAML with x-order set on definition properties, however
the new spec YAML is never actually used. This PR fixes this by
reloading g.Spec if PropertiesSpecOrder is set.