fix(flatten): fixed code generation with circular $ref AND expand option
Whenever generating code (server, client) from a spec with the
--with-flatten=expand option,
the spec document is unusable to resolve a circular $ref.

Although the root cause is most likely a bug in
go-openapi/spec/expander.go, an easier resolution path may be adopted
by cloning the loads.Document using a patched version of Pristine().

This PR requires https://github.com/go-openapi/loads/pull/51, as we
want to use Pristine() on a modified version of the spec inside the
loads.Document.

* fixes #2346

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>