fix(mangling): fixes name mangling in the presence of a special character
This PR fixes a special case when a "$" is inside a name AND we generate
polymorphic types.

In this case the unexported name was not properly generated (included a blank space).
More generally, swag mangling functions fail to properly handle this
extra blank space. Fortunately, ToGoName wasn't affected and this is why
this failure occurs in such rare circumstances.

* requires go-openapi/swag#78
* fixes #2821

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>