feat(validation): added --rooted-error-path to generate models
This new option to generate models allows to report more context
in validation errors for arrays and maps.

Validation errors used for those types used to be reported with an
empty path (e.g. " in body should have at least 1 items").

With this option enabled, generated models for arrays and maps
report the type (e.g. "[myArray] in body should have at least 1 items").

This flag acts as a feature flag: it is disabled by default with no
alteration of the existing generated code.

* fixes #2597

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>