fix(validations): fixed missing validation in embedded structs
* fixes #2604

This PR fixes the empty Validate methods generated when generating
embedded structs (that is, when an object redefines another object
with a $ref).

Since the actual need for validation is not known when hitting the
embedded struct, always assume a validation.

Exception: when the embedded type is interface{} or io.Reader, we
know that we should not call Validate.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>