fix(codegen): fixed calling order for WithAutoXOrder
This PR moves the call to WithAutoXOrder _before_ we validate and
flatten the spec for codegen purpose.

* fixes #2949

Other changes:
* layout: moved all the spec massaging to a dedicated source
* allowed WithAutoXOrder to report about errors rather than panicking

Overall, WithAutoXOrder is simplistic and could be improved a lot:
* supporting JSON
* supporting remote $refs, circular $refs
* removing the dependency to yaml.v2
* not producing a temporary file that hangs around
* etc

These improvements are beyond the scope of this PR, which is just to
fix responses. Perhaps a follow up could bring a more general solution to this problem.

Signed-off-by: Frédéric BIDON <fredbi@yahoo.com>