fix(codegen): fixed operation package mangling when tag is v1
* fixes #2866

This PR addresses that case when tags used to generate a package
to hold a group of operations are of the special form v[0-9]+.

This form is interpreted by go imports so that "x/v1" resolves as "x" in
the program.

To fix the problem, "v[nnn]" package names are renamed "version[nnn]".

Signed-off-by: Frédéric BIDON <fredbi@yahoo.com>