Reintroduced codeql and Open SSF score card
* The main disagreement on #3049 came from dependabot
* codeql scan and score card are thought as desirable additions

dependabot triggered a lot of PRs and overall, generates a lot of
extra work for the maintainers.

We need extra care and testing before introducing a dependency update
bot that understands our dependencies, may be coupled with auto
approval and auto merge, and may skip long running integration tests
when they are irrelevant. This work is deferred to a forthcoming PR.

The pinning of build dependencies (github actions, docker base images)
is reverted until we set out for a proper update process to
automatically update them with an appropriate frequency.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>