generator/templates: remove redundant warning for generated files
Commit f8fdd34ee024def2304ce2417b6bab6046b8cb02 added a "generated" comment
to all templates; this comment follows Go's standard notation, which is
detected by editors and linters to mark a file as generated. Those comments
make the custom comments further down the template redundant, as (most)
editors will already warn the user that they are editing a generated file.

This patch removes the redundant comment.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>