Relinted & refactored package cmd
This PR continues the relinting effort started by
PRs #3233, #3234 and #3236.

Build on top of #3236.

This time, we focus on the cmd package and in particular the
cmd/commands/diff which needed a bit of refactoring
(e.g. reducing the cognitive complexity of some function,
making struct initializers more readable etc).

All linting issues in this package are resolved with the current
linter configuration.

Contributes #3227

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>