Follow-up on PR #3195
The original PR proposes a new custom implementation of the go-imports
standard feature that is used by all-codegen tasks to format go code
with correct imports.

The performance improvement I could observe is significant: from 15% to 50%,
depending on the use case.

This commits adds the following:

* makes the adoption of the non-standard formatter an option, available
  from CLI
* document the new CLI option for generate commands
* relinted a few nitpicks, due to the addition of new linters since the
  original PR was first posted
* introduced this option in the CI codegen exerciser, which benefits
  from this performance improvement.

Overall, #3195 brings a significant improvement. However, I still feel
a bit concerned that pre-included default imports are only implemented
for the generated client. I don't really understand why we would have
missing imports only in that case. It is okay to leave that part to
some future PR.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>